backend/requirements.txt
alembic
bcrypt==4.0.1
passlib[bcrypt]
fastapi
uvicorn[standard]
psycopg2-binary
pydantic
python-jose[cryptography]
sqlalchemy
aiofiles
python-multipart
python-dotenv
email-validator
Jinja2
WeasyPrint


